# backend/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GestorPlus API",
        "description": "Documentación Swagger para la API de GestorPlus",
        "contact": {
            "name": "Soporte GestorPlus",
            "email": "soporte@tudominio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Servidor local"
        },
        {
            "url": "https://api.tudominio.com/",
            "description": "Servidor producción"
        }
    ],
    "paths": {
...
        "/vacaciones/solicitar": {
            "post": {
                "tags": [
                    "Vacaciones"
                ],
                "summary": "Solicitar vacaciones",
                "description": "Permite al usuario autenticado solicitar vacaciones.",
                "operationId": "6c6ccd3f1132c8a909997a1f5aab69e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fechaInicio",
                                    "fechaFin"
                                ],
                                "properties": {
                                    "fechaInicio": {
                                        "type": "string",
                                        "example": "2024-07-01"
                                    },
                                    "fechaFin": {
                                        "type": "string",
                                        "example": "2024-07-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacaciones solicitadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vacaciones solicitadas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al solicitar las vacaciones"
                    }
                }
            }
        },
        "/vacaciones/aceptar": {
            "patch": {
                "tags": [
                    "Vacaciones"
                ],
                "summary": "Aceptar solicitud de vacaciones",
                "description": "Permite aceptar una solicitud de vacaciones por idvacacion.",
                "operationId": "c6eacb53c48763aaa305880e5b44747c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idvacacion"
                                ],
                                "properties": {
                                    "idvacacion": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacación aceptada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "vacacionAceptada": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al aceptar la vacación"
                    }
                }
            }
        },
        "/vacaciones/rechazar": {
            "patch": {
                "tags": [
                    "Vacaciones"
                ],
                "summary": "Rechazar solicitud de vacaciones",
                "description": "Permite rechazar una solicitud de vacaciones por idvacacion.",
                "operationId": "5378b803355c4022981a16fd2469c39f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idvacacion"
                                ],
                                "properties": {
                                    "idvacacion": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacación rechazada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "vacacionRechazada": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al rechazar la vacación"
                    }
                }
            }
        },
...
    },
...
}